ReOrg / fix tests for utils move
